\name{firstbase}
\alias{firstbase}
\docType{data}
\title{
Rounding First Base.
}
\description{
The amount of time it took 22 baseball players to round first base for each of three methods of rounding.
}
\usage{data(firstbase)}
\format{
  A data frame with 22 observations on the following 3 variables.
  \describe{
    \item{\code{round.out}}{Time when using round out method.}
    \item{\code{narrow.angle}}{Time when using narrow angle method.}
    \item{\code{wide.angle}}{Time when using wide angle method.}
  }
}
\details{
Rounding methods are illustrated in Figure 7.1 of Hollander and Wolfe (1999).
}
\source{
Hollander, M. and Wolfe, D.A. (1999), \emph{Nonparametric Statistical Methods}, New York: Wiley.
}
\references{
Hollander, M. and Wolfe, D.A. (1999), \emph{Nonparametric Statistical Methods}, New York: Wiley.
}
\keyword{datasets}
