\name{hemorrhage}
\alias{hemorrhage}
\docType{data}
\title{
Hemorrhage data from Dupont.
}
\description{
Hemorrhage data from Dupont.
}
\usage{data(hemorrhage)}
\format{
  A data frame with 71 observations on the following 3 variables.
  \describe{
    \item{\code{genotype}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
    \item{\code{recur}}{a numeric vector}
  }
}
\references{
Dupont
}
\examples{
data(hemorrhage)
## maybe str(hemorrhage) ; plot(hemorrhage) ...
}
\keyword{datasets}
