\name{HoggQs}
\alias{Q1}
\alias{Q2}
\title{Hogg's Q1 and Q2.}
\description{ Q1 is a measure of skewness and Q2 is a measure of tail heaviness.  }
\usage{
Q1(z)
}
\arguments{
  \item{z}{n by 1 vector }
}
\details{
Used as selector statistics in adaptive schemes.
}
\references{
Hogg, R. McKean, J, Craig, A (2013) \emph{Introduction to Mathematical Statistics, 7th Ed.} Boston: Pearson.
}
\author{John Kloke <kloke@biostat.wisc.edu> }

\seealso{
\code{\link{hogg.test}}, ~~~
}
\keyword{nonparametrics}
