\name{internal}
\alias{lmean}
\alias{umean}
\title{ Internal Functions}
\description{ Internal functions not intended for general use.  
Used in calculation of Hogg's Qs.
}
\usage{
lmean(z, p)
}
\arguments{
  \item{z}{n by 1 vector}
  \item{p}{scalar}
}
\value{
Returns the calculated value as a numeric scalar.
}
\seealso{
\code{\link{hogg.test},\link{HoggsQs}}
}


\author{ John Kloke, Joseph McKean }
