\name{kancovarown}
\alias{kancovarown}
\title{ routine used in the ANCOVA table obtained by kancova}
\description{
routine used in making the display of the ANCOVA table obtained by kancova.
}
\usage{
kancovarown(vec)
}
\arguments{
  \item{vec}{vector to be labeled.}
}
\details{
Returns the labels.
}
\value{
  \item{nm}{vector of labels}
}
\references{
Kloke, J. and McKean, J.W. (2014), \emph{Nonparametric statistcal methods using R}, Boca Raton, FL: Chapman-Hall.
}
