\name{print}
\alias{print.hogg.test}
\alias{print.rank.test}
\alias{print.fkk.test}
\title{Internal print functions}
\description{Internal print functions} 
\usage{
\method{print}{hogg.test}(x, digits = max(5, .Options$digits - 2), ...)
\method{print}{rank.test}(x,...)
\method{print}{fkk.test}(x,...)
}
\arguments{
  \item{x}{Object to be printed.}
  \item{digits}{Number of digits to present.  Passed to print function.}
  \item{\dots}{ Additional arguments.}
}
\author{John Kloke, Joseph McKean}
